Subscribe to changes every N seconds.

+++ app/src/main/java/com/mathroda/dashcoin/feature_coins/presentation/coins_screen/components/CoinsItem.kt
@@ -129,7 +129,7 @@ fun CoinsItem(
 
 
                 val duration = 800
-                val formattedPrice = DecimalFormat("#,###,###.##").format(coinModel.price.toFloat())
+                val formattedPrice = DecimalFormat("#,###,###.##").format(coinModel.price)
                 AnimatedContent(targetState = "$currencySymbol $formattedPrice",
                     transitionSpec = {
                         slideInVertically(
